✨ Feat: 공통 훅으로 분리
feat: 공통 훅으로 분리
