+++ src/pages/QuestionForm/DatePage.tsx
@@ -1,4 +1,3 @@
-import { graniteEvent } from "@apps-in-toss/web-framework";
 import { adaptive } from "@toss/tds-colors";
 import {
 	FixedBottomCTA,
@@ -7,16 +6,16 @@ import {
 	Top,
 	WheelDatePicker,
 } from "@toss/tds-mobile";
-import { useEffect } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
 import { useSurvey } from "../../contexts/SurveyContext";
 import { pushGtmEvent } from "../../utils/gtm";
 import { formatQuestionNumber } from "../../utils/questionFactory";
+import { useQuestionBackHandler } from "./hooks/useQuestionBackHandler";
 import { useQuestionByType } from "./hooks/useQuestionByType";
 import { useCreateSurveyQuestion } from "./hooks/useQuestionMutations";
 
 export const DatePage = () => {
-	const { state, updateQuestion, deleteQuestion } = useSurvey();
+	const { state, updateQuestion } = useSurvey();
 	const { mutate: createSurveyQuestion } = useCreateSurveyQuestion();
 	const location = useLocation();
 	const navigate = useNavigate();
@@ -36,18 +35,7 @@ export const DatePage = () => {
 
 	const date = question?.date;
 
-	useEffect(() => {
-		const unsubscription = graniteEvent.addEventListener("backEvent", {
-			onEvent: () => {
-				if (!questionIdFromUrl && questionId) {
-					deleteQuestion(questionId);
-				}
-				navigate(-1);
-			},
-		});
-
-		return unsubscription;
-	}, [navigate, questionIdFromUrl, questionId, deleteQuestion]);
+	useQuestionBackHandler({ questionId, questionIdFromUrl });
 
 	const handleDateChange = (newDate: Date) => {
 		if (questionId) {
